Replace "bootstrap" with "update"
`update` performs the same actions as `bootstrap`, except that it also
stores the SHA of the file when it is installed.  This way, if the
current file and the stored SHA agree, then the file can be safely
replaced. If they do not agree, the user has modified the file, so it
will not be replaced without the 'install' flag.